Exception handling and swagger

+++ src/main/java/com/shop/CoffeeShop/service/ProductService.java
@@ -1,6 +1,9 @@
 package com.shop.CoffeeShop.service;
 
 
+import com.shop.CoffeeShop.Exceptions.NoSuchProductExistsException;
+import com.shop.CoffeeShop.Exceptions.NoSuchUserExistsException;
+import com.shop.CoffeeShop.Exceptions.ProductAlreadyExistsException;
 import com.shop.CoffeeShop.domain.Product;
 import com.shop.CoffeeShop.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
@@ -20,26 +23,46 @@ public class ProductService implements IProductService{
 
     @Override
     public Product addProduct(Product product) {
+        Product existingProduct = productRepository.findById(product.getProductId()).orElse(null);
+        if (existingProduct == null){
         log.info("add new product {} to the database", product.getName());
         return productRepository.save(product);
+        }
+        else {
+            throw new ProductAlreadyExistsException("Product is already exist");
+        }
     }
 
     public void deleteProduct(Long productId){
         Product product = productRepository.findById(productId).get();
+        if (product==null){
+            throw new NoSuchUserExistsException("Product Not Found");
+        }
+        else {
         log.info(" remove product {} from database", product.getName());
         productRepository.delete(product);
-    }
+    }}
 
     @Override
     public Product getProduct(String name) {
+        Product existingProduct = productRepository.findByName(name);
+        if (existingProduct == null){
+            throw new NoSuchProductExistsException("Product Not Found");
+        }
+        else {
         log.info("Fetching product {}", name);
         return productRepository.findByName(name);
-    }
+    }}
 
     public Product getProductById(Long productId) {
+        Product existingProduct = productRepository.findById(productId).orElse(null);
+        if (existingProduct == null){
+            throw new NoSuchProductExistsException("Product Not Found");
+        }
+        else {
         log.info("Fetching product with id {}", productId);
         return productRepository.findById(productId).get();
-    }
+    }}
 
     public List<Product> getAllProducts(){
         return productRepository.findAll();
